Refactor stats and dice pool into reusable components
